qa/qa_system/requirements.txt
sentence-transformers==2.2.2
numpy>=1.21.0
requests>=2.28.0
tqdm>=4.64.0
torch>=1.12.0
transformers>=4.21.0
scikit-learn>=1.1.0

# Agentic Framework 依赖
asyncio>=3.4.3
typing-extensions>=4.0.0

# Langchain 相关依赖
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-experimental>=0.0.50
langsmith>=0.0.80

# Ollama 集成依赖
ollama>=0.1.7

# 向量存储和检索依赖
langchain-faiss>=0.1.0

# 内存和会话管理
langchain-memory>=0.1.0

# Web前端依赖
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# 额外工具依赖
python-dotenv>=1.0.0
pydantic>=2.0.0 